requirements.txt
absl-py==0.9.0
pyfarmhash==0.2.2
matplotlib==3.0.3
numpy==1.16.4
scipy==1.2.1
seaborn==0.9.0
